AWS SNS 구독 확인 로그 추가

+++ src/main/java/com/heylocal/traveler/util/aws/S3ObjectNameFormatter.java
@@ -1,13 +1,12 @@
 package com.heylocal.traveler.util.aws;
 
-import com.heylocal.traveler.domain.travelon.opinion.OpinionImageContent;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
 
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 
-import static com.heylocal.traveler.domain.travelon.opinion.OpinionImageContent.*;
+import static com.heylocal.traveler.domain.travelon.opinion.OpinionImageContent.ImageContentType;
 
 /**
  * 포맷에 맞춰 S3의 Object 이름을 생성하는 클래스
@@ -30,7 +29,7 @@ public class S3ObjectNameFormatter {
    * @param objectIndex 이미지 배치 순서
    * @return
    */
-  public String getObjectNameOfOpinionImg(long travelOnId, long opinionId, ImageContentType imgType, long objectIndex) {
+  public String getObjectNameOfOpinionImg(long travelOnId, long opinionId, ImageContentType imgType, int objectIndex) {
     String objectName = OPINION_IMG_FORMAT;
 
     objectName = objectName.replace(TRAVEL_ON_ID_PARAM, String.valueOf(travelOnId));
